Move private definitions to `gui/private/common.h`
1
