+++ src/gui/render_scalers.cpp
@@ -8,8 +8,12 @@
 //just handle it all in one go, but this seems to work well enough for now
 
 #include "dosbox.h"
-#include "gui/render.h"
+
+#include "private/common.h"
+
 #include <cstring>
+
+#include "gui/render.h"
 
 uint8_t Scaler_Aspect[SCALER_MAXHEIGHT]        = {};
 uint16_t Scaler_ChangedLines[SCALER_MAXHEIGHT] = {};
